+++ custom_components/solarcharger/models/model_config.py
@@ -14,16 +14,24 @@ class ConfigValue:
     entity_id: str | None
     entity_value: Any | None
 
-    # # config_time -> entity_id
-    # config_entities: dict[str, str] = {}
+    # ----------------------------------------------------------------------------
+    def __repr__(self) -> str:
+        """Return string representation of config value."""
 
-    # # config_item -> value
-    # config_values: dict[str, Any] = {}
+        return f"{self.config_item}:(id={self.entity_id}, val={self.entity_value})"
 
 
+# ----------------------------------------------------------------------------
+# ----------------------------------------------------------------------------
 @dataclass
 class ConfigValueDict:
     """Dictionary of entity values."""
 
     config_item: str
     config_values: dict[str, ConfigValue]
+
+    # ----------------------------------------------------------------------------
+    def __repr__(self) -> str:
+        """Return string representation of config value."""
+
+        return f"{self.config_values.values()}"