+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Breaking change
 - Reminder: Please [delete then re-add the integration](https://github.com/flashg1/SolarCharger/wiki/Configuration#how-to-delete-and-re-add-the-solarcharger-integration) if upgrading from v0.5.4 or prior.
 ### Improvement
+- Reduced debug logging.
 - Refactored data models and reorganised directory structure.
 day.
 - Refactored max_charge_now to avoid pausing during the night and less likely to time-drift reducing charge speed during the day.